requirements.txt
Flask==3.0.0
firebase-admin==6.4.0
google-generativeai==0.8.3
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg[binary]==3.2.3
